+++ vpc/stack/main.tf
@@ -18,12 +18,13 @@ module "security_group" {
 }
 
 module "vpc_net" {
-  source                   = "../../vpc/networking"
-  iam_data                 = var.iam_data
-  monitor_data             = var.monitor_data
-  std_map                  = var.std_map
-  vpc_map                  = local.vpc_net_map
-  vpc_nat_multi_az_default = var.vpc_nat_multi_az_default
+  source                                     = "../../vpc/networking"
+  iam_data                                   = var.iam_data
+  monitor_data                               = var.monitor_data
+  std_map                                    = var.std_map
+  vpc_availability_zone_map_key_list_default = var.vpc_availability_zone_map_key_list_default
+  vpc_map                                    = local.vpc_net_map
+  vpc_nat_multi_az_default                   = var.vpc_nat_multi_az_default
 }
 
 module "vpc_peer" {